Balances HTML Build values
A lot of this game rides on frame-based activity (not great, I know) and will likely need to be updated. Play testing starting now on other devices to see how dramatic the change was. Just the difference between browser and editor leads me to believe that it will be drastic and each computer will have a different experience .